Added File Identifiers Setter

+++ src/main/java/com/techvisio/einstitution/beans/AdmissionDiscountDtl.java
@@ -3,7 +3,7 @@ package com.techvisio.einstitution.beans;
 public class AdmissionDiscountDtl {
 
 	private String fileNo;
-	private int feeHeadId;
+	private Long feeHeadId;
 	private double amount;
 	private float percent;
 	
@@ -17,11 +17,11 @@ public class AdmissionDiscountDtl {
 	}
 	
 	
-	public int getFeeHeadId() {
+	public Long getFeeHeadId() {
 		return feeHeadId;
 	}
 	
-	public void setFeeHeadId(int feeHeadId) {
+	public void setFeeHeadId(Long feeHeadId) {
 		this.feeHeadId = feeHeadId;
 	}
